+++ packages/caravan-clients/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @caravan/clients
 
+## 1.0.2
+
+### Patch Changes
+
+- [#412](https://github.com/caravan-bitcoin/caravan/pull/412) [`ee9e1dd`](https://github.com/caravan-bitcoin/caravan/commit/ee9e1dd3c5b99545b9007fc59239227fefe9a8a4) Thanks [@Shadouts](https://github.com/Shadouts)! - update some packages
+
+- Updated dependencies [[`ee9e1dd`](https://github.com/caravan-bitcoin/caravan/commit/ee9e1dd3c5b99545b9007fc59239227fefe9a8a4)]:
+  - @caravan/bitcoin@0.4.2
+
+## 1.0.1
+
+### Patch Changes
+
+- [#352](https://github.com/caravan-bitcoin/caravan/pull/352) [`4d5fd81`](https://github.com/caravan-bitcoin/caravan/commit/4d5fd8136126dc98d0baa20dd445ce80ae0a4910) Thanks [@ice-009](https://github.com/ice-009)! - Added the transaction history tab in the coordinator and some minor changes in the caravan-clients to meet that need
+
 ## 1.0.0
 
 ### Major Changes